added separate files back into index.js
